==> code/03_enriching_data.ipynb <==
# data enrichment

# importing 
import pandas as pd

# load full XETRA dataset
df_full = pd.read_csv("../data/07_17_to_01_21_full.csv")

df_full.head()

# load ticker and industry/sector mapping
df_mapping = pd.read_csv("../data/ticker_sector_industry_mapping.csv")
df_mapping

# proof of conpect using first 10 rows
df_test_data = df_full[0:10]
df_test_map = df_mapping[:10]

# rename ticker column to original name in order to allow for left join
df_test_map.rename(columns = {"ticker": "Mnemonic"}, inplace = True)

df_test_data.merge(df_test_map, how = "left", on="Mnemonic")

# rename column of full mapping
df_mapping.rename(columns = {"ticker": "Mnemonic"}, inplace = True)

# merge full datas
df_enriched = df_full.merge(df_mapping, how = "left", on="Mnemonic")

# inspect NaNs
df_enriched.isna().sum()

# quick look into most popular sectors
df_enriched["sector"].value_counts()

# quick look into most popular industries
df_enriched["industry"].value_counts()

# save enriched dataset
df_enriched.to_csv("../data/07_17_to_01_21_full_enriched.csv", index = None)

# test loading enriched dataset
df_enriched_loaded = pd.read_csv("../data/07_17_to_01_21_full_enriched.csv")

# inspect loaded df
df_enriched_loaded.head()
