==> code/04_Retrieve_shares_outstanding_DAX_MDax_SDax.ipynb <==
# Introduce imports which we need
import pandas as pd
import numpy as np
from yahoofinancials import YahooFinancials

# Retrieve Shares Outstanding for Companies in Dax, MDax, SDax, TecDax

# Read DataFrame
df_dax_all = pd.read_excel('.../data/German_Indices.xlsx')
# Create new column to store info about shares outstanding
df_dax_all['Shares_Outstanding'] = np.zeros(len(df_dax_all))

# Retrieve info about shares outstanding
ticker_symbols = df_dax_all['Mnemonic']
dict_shares_outs = {}
for i in list(range(0, len(ticker_symbols))):
    ticker = ticker_symbols[i]+str('.F')
    yahoo_financials = YahooFinancials(ticker)
    shares_outstanding = yahoo_financials.get_num_shares_outstanding()
    dict_shares_outs[ticker_symbols[i]] = shares_outstanding

# Create list out of shares_outstanding to insert it into the dataframe
list_shares_outs = list(dict_shares_outs.values())

# Insert Shares Outstanding into DataFrame
df_dax_all['Shares_Outstanding'] = list_shares_outs

# Save file
df_dax_all.to_excel('.../data/Shares_Outstanding.xlsx')
