==> code/05_Enrich_Index_Shares Outstanding_Country.ipynb <==
# Introduce imports which we need
import pandas as pd
import numpy as np

# Read csv-file
df_all = pd.read_csv('.../data/07_17_to_01_21_full_enriched.csv')

# Create column for the country code
df_all['ISO_Code'] = df_all['ISIN'].str[:2]

# Read excel-file with index information and shares outstanding
df_index_shares = pd.read_excel('.../data/Shares_Outstanding.xlsx')

# merge information about indices and shares outstanding
df_all = df_all.merge(df_index_shares, how = "left", on="Mnemonic")

# Read excel-file with country information
df_countries = pd.read_excel('.../data/ISO_Country_Codes.xlsx')

# merge information about contries
df_all = df_all.merge(df_countries, how = "left", on="ISO_Code")

# Save DataFrame to csv-file
df_all.to_csv('.../data/07_17_to_01_21_final_dataset.csv')
